mades some new changes regarding fundtransfer

+++ src/main/java/com/bank/dao/LoanDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bank.model.Loan;
 
-public interface LoanDAO extends JpaRepository<Loan, Integer> {
+public interface LoanDAO extends JpaRepository<Loan, Long> {
 
 }
